Update compile_path_example.py
Testing with your version showed, that the root folder `Datacenters` has been added to the path, which is wrong. So now, if we bail out of the loop because we reached the top, we need to remove this root folder `Datacenters` (or `Datencenter` as it is called on our server). This should only happen for "datacenters" because we also break the loop, as soon as this object is not a folder.

So given a datacenter, this now returns the full path as is. For a vm, datastore or network, this returns the full path `relative` to its `datacenter` this object is in.

I'm not sure if this behaviour is also correct (or relevant) for resource pools.

Another difference to the simulator on my vcenter 5.5. When I'm on the top `datacenters` folder and calling `thisobj.parent` it returns `None`. On the simulator, this directly returns the `IndexError`

```
'vim.Folder:group-d1'
(Pdb) n
> /ansible/playbooks/get_folder_path(54)compile_folder_path_for_object()
-> if isinstance(thisobj, vim.Folder):
(Pdb) 
> /ansible/playbooks/get_folder_path(55)compile_folder_path_for_object()
-> paths.append(thisobj.name)
(Pdb) 
> /ansible/playbooks/get_folder_path(45)compile_folder_path_for_object()
-> while True:
(Pdb) 
> /ansible/playbooks/get_folder_path(46)compile_folder_path_for_object()
-> try:
(Pdb) 
> /ansible/playbooks/get_folder_path(47)compile_folder_path_for_object()
-> thisobj = thisobj.parent
(Pdb) 
> /ansible/playbooks/get_folder_path(48)compile_folder_path_for_object()
-> if thisobj is None:
(Pdb) p thisobj
None
(Pdb) 
...
```

VCSIM

```
> /ansible/playbooks/get_folder_path(51)compile_folder_path_for_object()
-> if isinstance(thisobj, vim.Folder):
(Pdb) p thisobj
'vim.Folder:group-d1'
(Pdb) n
> /ansible/playbooks/get_folder_path(52)compile_folder_path_for_object()
-> paths.append(thisobj.name)
(Pdb) 
> /ansible/playbooks/get_folder_path(46)compile_folder_path_for_object()
-> try:
(Pdb) 
> /ansible/playbooks/get_folder_path(47)compile_folder_path_for_object()
-> thisobj = thisobj.parent
(Pdb) 
IndexError: list index out of range
> /ansible/playbooks/get_folder_path(47)compile_folder_path_for_object()
-> thisobj = thisobj.parent
(Pdb) 
>>> 
```